# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is gopa-server. Use these APIs to manage different groups, roles of GOPA",
        "title": "Gopa-server Swagger Documentation",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/projectResource/update": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Update the permission of a group with its related role to a resource",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projectResource"
                ],
                "summary": "ProjectResourceUpdate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdatedProjectResourceForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/rego/update": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Update the content of a rego file by given path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rego"
                ],
                "summary": "RegoUpdate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdatedRegoDocumentForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/update": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Update the group or role of a given user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "UserUpdate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdatePermissionForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.UpdatePermissionForm": {
            "type": "object",
            "properties": {
                "project": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.UpdatedProjectResourceForm": {
            "type": "object",
            "properties": {
                "project_name": {
                    "type": "string"
                },
                "resource_name": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
        "api.UpdatedRegoDocumentForm": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}